@@ -13,14 +13,14 @@ import 'package:intl/intl.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class BookAppointment extends StatefulWidget {
-  final String drName;
-  final String category;
-  final String fee;
-  final String image;
-  final String drUsername;
+  final String? drName;
+  final String? category;
+  final String? fee;
+  final String? image;
+  final String? drUsername;
 
   const BookAppointment(
-      {Key key,
+      {Key? key,
       this.drName,
       this.drUsername,
       this.fee,
@@ -34,11 +34,11 @@ class BookAppointment extends StatefulWidget {
 
 class _BookAppointmentState extends State<BookAppointment> {
   int selected = 0;
-  int timeSelect;
-  bool isChecked = false;
+  int? timeSelect;
+  bool? isChecked = false;
 
-  List<TimeSlot> timeSlots;
-  TimeSlot timeSlot;
+  List<TimeSlot>? timeSlots;
+  TimeSlot? timeSlot;
 
   bool isTaped = true;
   String buttonText = "Confirm Appointment";
@@ -53,7 +53,7 @@ class _BookAppointmentState extends State<BookAppointment> {
       ),
       body: Container(
         padding: const EdgeInsets.only(left: 16, right: 16, top: 16, bottom: 10),
-        child: timeSlots.isNotEmpty
+        child: timeSlots!.isNotEmpty
             ? view()
             : AppointmentShimmer(),
       ),
@@ -78,9 +78,9 @@ class _BookAppointmentState extends State<BookAppointment> {
         "&DoctorUsername=${widget.drUsername}&DaysCount=4");
 
     if (response != "404") {
-      timeSlots = timeSlotModelFromJson(response).response.response;
-      if (timeSlots.isNotEmpty) {
-        timeSlot = timeSlots[0];
+      timeSlots = timeSlotModelFromJson(response).response!.response;
+      if (timeSlots!.isNotEmpty) {
+        timeSlot = timeSlots![0];
         if (!mounted) return;
         setState(() {});
       } else {
@@ -115,7 +115,7 @@ class _BookAppointmentState extends State<BookAppointment> {
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
                     Text(
-                      widget.drName,
+                      widget.drName!,
                       style: const TextStyle(fontWeight: FontWeight.bold, fontSize: 22),
                       maxLines: 2,
                     ),
@@ -123,7 +123,7 @@ class _BookAppointmentState extends State<BookAppointment> {
                       height: 8,
                     ),
                     Text(
-                      widget.category,
+                      widget.category!,
                       style: const TextStyle(
                           fontWeight: FontWeight.bold,
                           fontSize: 18,
@@ -138,7 +138,7 @@ class _BookAppointmentState extends State<BookAppointment> {
                           color: MyColors.accent,
                           borderRadius: BorderRadius.circular(8)),
                       child: Text(
-                        widget.fee,
+                        widget.fee!,
                         style: const TextStyle(color: Colors.white),
                       ),
                     )
@@ -156,8 +156,8 @@ class _BookAppointmentState extends State<BookAppointment> {
           child: ListView.builder(
             itemBuilder: (BuildContext context, int index) {
               bool select = selected == index ? true : false;
-              TimeSlot date = timeSlots[index];
-              DateTime dateObj = DateFormat("MM/dd/yyyy").parse(date.date);
+              TimeSlot date = timeSlots![index];
+              DateTime dateObj = DateFormat("MM/dd/yyyy").parse(date.date!);
               DateTime nowDateTime =
                   DateTime(dateObj.year, dateObj.month, dateObj.day);
 
@@ -218,7 +218,7 @@ class _BookAppointmentState extends State<BookAppointment> {
                 ),
               );
             },
-            itemCount: timeSlots.length,
+            itemCount: timeSlots!.length,
             scrollDirection: Axis.horizontal,
           ),
         ),
@@ -234,12 +234,12 @@ class _BookAppointmentState extends State<BookAppointment> {
               crossAxisSpacing: 8,
               mainAxisSpacing: 8,
               scrollDirection: Axis.vertical,
-              children: timeSlot.timeSlots.map((e) {
-                int index = timeSlot.timeSlots.indexOf(e);
+              children: timeSlot!.timeSlots!.map((e) {
+                int index = timeSlot!.timeSlots!.indexOf(e);
                 bool timeSelected = timeSelect == index ? true : false;
                 return GestureDetector(
                   onTap: () {
-                    timeSelect = timeSlot.timeSlots.indexOf(e);
+                    timeSelect = timeSlot!.timeSlots!.indexOf(e);
                     setState(() {});
                   },
                   child: Container(
@@ -270,7 +270,7 @@ class _BookAppointmentState extends State<BookAppointment> {
         GestureDetector(
           onTap: (){
             setState(() {
-              isChecked = !isChecked;
+              isChecked = !isChecked!;
             });
           },
           child: Row(
@@ -347,12 +347,12 @@ class _BookAppointmentState extends State<BookAppointment> {
   void bookAppointment() async {
     disableButton();
     SharedPreferences preferences = await SharedPreferences.getInstance();
-    String username = preferences.getString(Keys.username);
-    String name = preferences.getString(Keys.name);
+    String username = preferences.getString(Keys.username)!;
+    String name = preferences.getString(Keys.name)!;
 
-    String time = timeSlot.timeSlots[timeSelect];
-    String newDate = timeSlot.date + " " + time;
-    String checkUp = isChecked ? "Online" : "Regular Checkup";
+    String time = timeSlot!.timeSlots![timeSelect!];
+    String newDate = timeSlot!.date! + " " + time;
+    String checkUp = isChecked! ? "Online" : "Regular Checkup";
 
     String values = "&DoctorUsername=${widget.drUsername}" "&Location=${Keys.locationId}" "&patname=" +
         name +
