@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 
 class DiagnosticsHome extends StatelessWidget {
-  const DiagnosticsHome({Key key}) : super(key: key);
+  const DiagnosticsHome({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
