@@ -3,7 +3,7 @@ import 'package:amc/Widgets/loading_dialog.dart';
 import 'package:flutter/material.dart';
 
 class MyLabTestReports extends StatefulWidget {
-  const MyLabTestReports({Key key}) : super(key: key);
+  const MyLabTestReports({Key? key}) : super(key: key);
 
   @override
   _MyLabTestReportsState createState() => _MyLabTestReportsState();
