@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_inappwebview/flutter_inappwebview.dart';
 
 class PrescriptionWebView extends StatefulWidget {
-  final String url;
-  final String title;
+  final String? url;
+  final String? title;
 
-  const PrescriptionWebView(this.url, this.title, {Key key}) : super(key: key);
+  const PrescriptionWebView(this.url, this.title, {Key? key}) : super(key: key);
 
   @override
   _PrescriptionWebViewState createState() => _PrescriptionWebViewState(url);
@@ -13,8 +13,8 @@ class PrescriptionWebView extends StatefulWidget {
 
 class _PrescriptionWebViewState extends State<PrescriptionWebView> {
   double progress = 0;
-  String url;
-  InAppWebViewController webView;
+  String? url;
+  InAppWebViewController? webView;
 
   _PrescriptionWebViewState(this.url);
 
@@ -22,7 +22,7 @@ class _PrescriptionWebViewState extends State<PrescriptionWebView> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text(widget.title),
+        title: Text(widget.title!),
       ),
       body: Column(
         children: [
@@ -33,19 +33,19 @@ class _PrescriptionWebViewState extends State<PrescriptionWebView> {
             child: Container(
               padding: const EdgeInsets.all(8),
               child: InAppWebView(
-                initialUrlRequest: URLRequest(url: Uri.parse(url)),
+                initialUrlRequest: URLRequest(url: Uri.parse(url!)),
                 initialOptions: InAppWebViewGroupOptions(
                     crossPlatform: InAppWebViewOptions(
                         supportZoom: true, cacheEnabled: true)),
                 onWebViewCreated: (InAppWebViewController controller) {
                   webView = controller;
                 },
-                onLoadStart: (InAppWebViewController controller, Uri url) {
+                onLoadStart: (InAppWebViewController controller, Uri? url) {
                   setState(() {
                     this.url = url.toString();
                   });
                 },
-                onLoadStop: (InAppWebViewController controller, Uri url) async {
+                onLoadStop: (InAppWebViewController controller, Uri? url) async {
                   setState(() {
                     this.url = url.toString();
                   });
