@@ -8,7 +8,7 @@ import 'package:shared_preferences/shared_preferences.dart';
 import 'Styles/Keys.dart';
 
 class Splash extends StatefulWidget {
-  const Splash({Key key}) : super(key: key);
+  const Splash({Key? key}) : super(key: key);
 
   @override
   _SplashState createState() => _SplashState();
