@@ -2,13 +2,13 @@ import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 
 class NetWorkImage extends StatelessWidget {
-  final String imagePath;
-  final String placeHolder;
-  final double width;
-  final double height;
+  final String? imagePath;
+  final String? placeHolder;
+  final double? width;
+  final double? height;
 
   const NetWorkImage(
-      {Key key, this.imagePath, this.placeHolder, this.width, this.height})
+      {Key? key, this.imagePath, this.placeHolder, this.width, this.height})
       : super(key: key);
 
   @override
@@ -31,31 +31,31 @@ class NetWorkImage extends StatelessWidget {
       width: width,
       height: height,
       fit: BoxFit.cover,
-      placeholder: (context, url) => Image.asset(placeHolder),
-      errorWidget: (context, url, error) => Image.asset(placeHolder),
+      placeholder: (context, url) => Image.asset(placeHolder!),
+      errorWidget: (context, url, error) => Image.asset(placeHolder!),
     );
   }
 }
 
 class CacheImage extends StatelessWidget {
-  final String imagePath;
-  final String placeHolder;
-  final double width;
-  final double height;
+  final String? imagePath;
+  final String? placeHolder;
+  final double? width;
+  final double? height;
 
   const CacheImage(
-      {Key key, this.imagePath, this.placeHolder, this.width, this.height})
+      {Key? key, this.imagePath, this.placeHolder, this.width, this.height})
       : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return CachedNetworkImage(
-      imageUrl: imagePath,
+      imageUrl: imagePath!,
       width: width,
       height: height,
       fit: BoxFit.fill,
-      placeholder: (context, url) => Image.asset(placeHolder),
-      errorWidget: (context, url, error) => Image.asset(placeHolder),
+      placeholder: (context, url) => Image.asset(placeHolder!),
+      errorWidget: (context, url, error) => Image.asset(placeHolder!),
     );
   }
 }
