@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class LazyLoader extends StatelessWidget {
-  const LazyLoader({Key key}) : super(key: key);
+  const LazyLoader({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
