@@ -5,18 +5,18 @@ import 'package:path_provider/path_provider.dart';
 import 'package:sqflite/sqflite.dart';
 
 class DatabaseManager {
-  static DatabaseManager _databaseManger;
-  static Database _database;
+  static DatabaseManager? _databaseManger;
+  static late Database _database;
 
   DatabaseManager._createInstance();
 
   factory DatabaseManager() {
     _databaseManger ??= DatabaseManager._createInstance();
-    return _databaseManger;
+    return _databaseManger!;
   }
 
   Future<Database> get database async {
-    _database ??= await initializeDatabase();
+    _database = await initializeDatabase();
     return _database;
   }
 
@@ -30,13 +30,11 @@ class DatabaseManager {
 
   void _createDb(Database db, int newVersion) async {
     await db.execute(Keys.notificationSchema);
-
   }
 
   Future<List<Map<String, dynamic>>> getMapList(String table) async {
     Database db = await database;
-    return await db.query(table,
-        orderBy: '${Keys.ID} DESC');
+    return await db.query(table, orderBy: '${Keys.ID} DESC');
   }
 
   Future<int> insert(String table, dynamic model) async {
@@ -49,15 +47,15 @@ class DatabaseManager {
     return await db.delete(table);
   }
 
-  Future<int> getCount(String table) async {
+  Future<int?> getCount(String table) async {
     Database db = await database;
 
-    List<Map<String, dynamic>> list = await db.rawQuery('SELECT COUNT (*) FROM $table');
+    List<Map<String, dynamic>> list =
+        await db.rawQuery('SELECT COUNT (*) FROM $table');
     return Sqflite.firstIntValue(list);
   }
 
   Future<List<NotificationModel>> getNotification() async {
-
     var noteMapList = await getMapList(Keys.notification);
     int count = noteMapList.length;
 
@@ -69,5 +67,4 @@ class DatabaseManager {
 
     return list;
   }
-
 }
