@@ -16,8 +16,8 @@ class SearchLabTestModel {
     this.response,
   });
 
-  Meta meta;
-  SearchLabTestModelResponse response;
+  Meta? meta;
+  SearchLabTestModelResponse? response;
 
   factory SearchLabTestModel.fromJson(Map<String, dynamic> json) => SearchLabTestModel(
     meta: Meta.fromJson(json["Meta"]),
@@ -25,8 +25,8 @@ class SearchLabTestModel {
   );
 
   Map<String, dynamic> toJson() => {
-    "Meta": meta.toJson(),
-    "Response": response.toJson(),
+    "Meta": meta!.toJson(),
+    "Response": response!.toJson(),
   };
 }
 
@@ -35,14 +35,14 @@ class SearchLabTestModelResponse {
     this.response,
   });
 
-  ResponseResponse response;
+  ResponseResponse? response;
 
   factory SearchLabTestModelResponse.fromJson(Map<String, dynamic> json) => SearchLabTestModelResponse(
     response: ResponseResponse.fromJson(json["Response"]),
   );
 
   Map<String, dynamic> toJson() => {
-    "Response": response.toJson(),
+    "Response": response!.toJson(),
   };
 }
 
@@ -52,8 +52,8 @@ class ResponseResponse {
     this.testsList,
   });
 
-  int totalItems;
-  List<Test> testsList;
+  int? totalItems;
+  List<Test>? testsList;
 
   factory ResponseResponse.fromJson(Map<String, dynamic> json) => ResponseResponse(
     totalItems: json["TotalItems"],
@@ -62,7 +62,7 @@ class ResponseResponse {
 
   Map<String, dynamic> toJson() => {
     "TotalItems": totalItems,
-    "TestsList": List<dynamic>.from(testsList.map((x) => x.toJson())),
+    "TestsList": List<dynamic>.from(testsList!.map((x) => x.toJson())),
   };
 }
 
@@ -77,13 +77,13 @@ class Test {
     this.fee,
   });
 
-  String id;
-  String testId;
-  String testName;
-  String sampleRequired;
-  String description;
-  String views;
-  String fee;
+  String? id;
+  String? testId;
+  String? testName;
+  String? sampleRequired;
+  String? description;
+  String? views;
+  String? fee;
 
   factory Test.fromJson(Map<String, dynamic> json) => Test(
     id: json["ID"],
