@@ -3,10 +3,10 @@ import 'package:flutter/material.dart';
 class MyColors{
   static const Color primary = Color(0xFF1D2093);
   static const Color accent = Color(0xFFFF9800);
-  static const Color primary_dark = Color(0xFF303F9F);
-  static const Color primary_light = Color(0xFFC5CAE9);
-  static const Color primary_text = Color(0xFF212121);
-  static const Color secondary_text = Color(0xFF757575);
+  static const Color primaryDark = Color(0xFF303F9F);
+  static const Color primaryLight = Color(0xFFC5CAE9);
+  static const Color primaryText = Color(0xFF212121);
+  static const Color secondaryText = Color(0xFF757575);
   static const Color icons = Color(0xFFFFFFFF);
   static const Color divider = Color(0xFFBDBDBD);
   static const highLight = Color(0xFFE0E0E0);
